+++ packages/spacecat-shared-data-access/src/models/site/config.js
@@ -16,6 +16,7 @@ export const IMPORT_TYPES = {
   ORGANIC_KEYWORDS: 'organic-keywords',
   ORGANIC_TRAFFIC: 'organic-traffic',
   TOP_PAGES: 'top-pages',
+  ALL_TRAFFIC: 'all-traffic',
 };
 
 export const IMPORT_DESTINATIONS = {
@@ -25,28 +26,40 @@ export const IMPORT_DESTINATIONS = {
 export const IMPORT_SOURCES = {
   AHREFS: 'ahrefs',
   GSC: 'google',
+  RUM: 'rum',
 };
 
 const IMPORT_BASE_KEYS = {
   destinations: Joi.array().items(Joi.string().valid(IMPORT_DESTINATIONS.DEFAULT)).required(),
   sources: Joi.array().items(Joi.string().valid(...Object.values(IMPORT_SOURCES))).required(),
-  enabled: Joi.boolean().required().default(true),
+  // not required for now due backward compatibility
+  enabled: Joi.boolean().default(true),
+  url: Joi.string().uri().optional(), // optional url to override
 };
 
 export const IMPORT_TYPE_SCHEMAS = {
   [IMPORT_TYPES.ORGANIC_KEYWORDS]: Joi.object({
     type: Joi.string().valid(IMPORT_TYPES.ORGANIC_KEYWORDS).required(),
     ...IMPORT_BASE_KEYS,
-    pageUrl: Joi.string().uri(),
+    geo: Joi.string().optional(),
+    limit: Joi.number().integer().min(1).max(100)
+      .optional(),
+    pageUrl: Joi.string().uri().optional(),
   }),
   [IMPORT_TYPES.ORGANIC_TRAFFIC]: Joi.object({
     type: Joi.string().valid(IMPORT_TYPES.ORGANIC_TRAFFIC).required(),
     ...IMPORT_BASE_KEYS,
   }),
+  [IMPORT_TYPES.ALL_TRAFFIC]: Joi.object({
+    type: Joi.string().valid(IMPORT_TYPES.ALL_TRAFFIC).required(),
+    ...IMPORT_BASE_KEYS,
+  }),
   [IMPORT_TYPES.TOP_PAGES]: Joi.object({
     type: Joi.string().valid(IMPORT_TYPES.TOP_PAGES).required(),
     ...IMPORT_BASE_KEYS,
-    geo: Joi.string(),
+    geo: Joi.string().optional(),
+    limit: Joi.number().integer().min(1).max(2000)
+      .optional(),
   }),
 };
 
@@ -61,6 +74,12 @@ export const DEFAULT_IMPORT_CONFIGS = {
     type: 'organic-traffic',
     destinations: ['default'],
     sources: ['ahrefs'],
+    enabled: true,
+  },
+  'all-traffic': {
+    type: 'all-traffic',
+    destinations: ['default'],
+    sources: ['rum'],
     enabled: true,
   },
   'top-pages': {
@@ -81,8 +100,12 @@ export const configSchema = Joi.object({
   imports: Joi.array().items(
     Joi.alternatives().try(...Object.values(IMPORT_TYPE_SCHEMAS)),
   ),
+  brandConfig: Joi.object({
+    brandId: Joi.string().required(),
+  }).optional(),
   fetchConfig: Joi.object({
     headers: Joi.object().pattern(Joi.string(), Joi.string()),
+    overrideBaseURL: Joi.string().uri().optional(),
   }).optional(),
   handlers: Joi.object().pattern(Joi.string(), Joi.object({
     mentions: Joi.object().pattern(Joi.string(), Joi.array().items(Joi.string())),
@@ -100,6 +123,8 @@ export const configSchema = Joi.object({
       name: Joi.string(),
       pattern: Joi.string(),
     })).optional(),
+    movingAvgThreshold: Joi.number().min(1).optional(),
+    percentageChangeThreshold: Joi.number().min(1).optional(),
     latestMetrics: Joi.object({
       pageViewsChange: Joi.number(),
       ctrChange: Joi.number(),
@@ -118,7 +143,7 @@ export function validateConfiguration(config) {
   const { error, value } = configSchema.validate(config);
 
   if (error) {
-    throw new Error(`Configuration validation error: ${error.message}`);
+    throw new Error(`Configuration validation error: ${error.message}`, { cause: error });
   }
 
   return value; // Validated and sanitized configuration
@@ -142,6 +167,7 @@ export const Config = (data = {}) => {
   self.getGroupedURLs = (type) => state?.handlers?.[type]?.groupedURLs;
   self.getLatestMetrics = (type) => state?.handlers?.[type]?.latestMetrics;
   self.getFetchConfig = () => state?.fetchConfig;
+  self.getBrandConfig = () => state?.brandConfig;
 
   self.updateSlackConfig = (channel, workspace, invitedUserCount) => {
     state.slack = {
@@ -198,6 +224,10 @@ export const Config = (data = {}) => {
     state.fetchConfig = fetchConfig;
   };
 
+  self.updateBrandConfig = (brandConfig) => {
+    state.brandConfig = brandConfig;
+  };
+
   self.enableImport = (type, config = {}) => {
     if (!IMPORT_TYPE_SCHEMAS[type]) {
       throw new Error(`Unknown import type: ${type}`);
@@ -249,4 +279,5 @@ Config.toDynamoItem = (config) => ({
   handlers: config.getHandlers(),
   imports: config.getImports(),
   fetchConfig: config.getFetchConfig(),
+  brandConfig: config.getBrandConfig(),
 });