+++ packages/spacecat-shared-data-access/src/models/suggestion/suggestion.model.js
@@ -28,6 +28,7 @@ class Suggestion extends BaseModel {
     SKIPPED: 'SKIPPED',
     FIXED: 'FIXED',
     ERROR: 'ERROR',
+    OUTDATED: 'OUTDATED',
   };
 
   static TYPES = {